teardown_send_job_report/requirements.txt
cloudshell-orch-core
cloudshell-sandbox-reporter